When we free and kill a process also wait for it
Otherwise we leave zombies laying around, which are a sanitary hazard
at best, and very confusing at worst, because it is not clear whether
there is a problem with the process or with the monitoring itself. Not
reaping zombies properly might also run afoul of system limits.

This happens commonly when timeouts trigger, as the process structure
is freed but nothing waits for the pid, and then the pid is lost.