Add Box_setColumn() for setting specific column value and Box_printRow() to print the row.
If some column value was not set, padding is used in the given column.

The object is kept, as it allows to keep the Box_T definition private in Box.c (using struct Box_T would require exporting it to Box.h)

The API supports value truncation (cut the value) or wrap (expand the column across multiple rows for long values).

The column can be either left or right aligned.

The table is defined in Box_new() parameters, then the Box_setColumn() can be used with plain format string ... the Box API internally handles the alignment, wrapping/truncating, etc.

Example:

        Box_T t = Box_new(res->outputbuffer, 2, (BoxColumn_T []){
                        {.name = "Column 1", .width = 30, .wrap = false, .align = BoxAlign_Left},
                        {.name = "Column 2", .width = 30, .wrap = true,  .align = BoxAlign_Right},
                  }, true);
        for (int i = 0; i < 10; i++) {
                Box_setColumn(t, 1, "%d", i);
                Box_setColumn(t, 2, "%s", "test");
                Box_printRow(t);
        }
        Box_free(&t);