Fixed: Some linux versions logged "Skipping redundant host 'localhost'" warning in the case that the /etc/hosts contained localhost entry for both "127.0.0.1" (IPv4 loopback) and "::1" (IPv6 loopback).
How to reproduce:

/etc/hosts:
        127.0.0.1 localhost
        ::1       localhost

Added debug code to Engine_addHostAllow:

        DEBUG: Engine_addHostAllow: adding pattern=localhost
        DEBUG: Engine_addHostAllow: getaddrinfo: #result1: ai_addr=127.0.0.1, ai_family=2, ai_flags=0, ai_addrlen=16, ai_socktype=1, ai_protocol=6, ai_canonname=(null), next=0xfa92a0
        DEBUG: Engine_addHostAllow: getaddrinfo: #result2: ai_addr=127.0.0.1, ai_family=2, ai_flags=0, ai_addrlen=16, ai_socktype=1, ai_protocol=6, ai_canonname=(null), next=(nil)

Although there is just one localhost entry for IPv4, the getaddrinfo() gets two results with the same IP. When the IPv6 localhost is removed, the problem doesn't show up.

We use getaddrinfo() hints with family set to AF_INET, as we're interested in IPv4 only currently. When this hint is removed, the duplicate doesn't show up.

The duplicate also doesn't show up on Ubuntu 14.04 nor Debian 4 ... the problem seems to be glibc bug related to newer linux versions only when getaddrinfo() hints family is set to AF_INET (IPv4).

Removal of the address family removed the duplicate - we're testing the result family when processing the getaddrinfo result set anyway.